third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-request-basics.html

<!DOCTYPE html>
<title>Credential Manager: request() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
function stubResolverUndefinedChecker(c) {
    assert_equals(c, undefined);
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("request() should not reject, but did: " + reason.name);
}

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request().then(
            t.step_func(stubResolverUndefinedChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request({}).");
    t.step(function () {
        navigator.credentials.request({}).then(
            t.step_func(stubResolverUndefinedChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            federations: [ 'https://example.com/' ]
        }).then(
            t.step_func(stubResolverUndefinedChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            federations: [ 'https://example.com/' ],
            zeroClickOnly: true
        }).then(
            t.step_func(stubResolverUndefinedChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            zeroClickOnly: true
        }).then(
            t.step_func(stubResolverUndefinedChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            notValid: 'yay!'
        }).then(
            t.step_func(stubResolverUndefinedChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function () {
    var id = "id";
    var name = "name";
    var avatar = "http://example.com/";
    var password = "pencil";

    function stubResolverChecker(c) {
        verifyInterface('LocalCredential', c, {
            id: 'string',
            name: 'string',
            avatarURL: 'string',
            password: 'string'
        });

        assert_equals(c.id, id);
        assert_equals(c.name, name);
        assert_equals(c.avatarURL, avatar);
        assert_equals(c.password, password);
        this.done();
    }

    var t = async_test("Verify that the mock returns the values we give it.");
    t.step(function() {
        if (window.testRunner)
            testRunner.addMockCredentialManagerResponse(id, name, avatar, password);
        navigator.credentials.request().then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });   
}());
</script>
